+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug preventing post story saves to go endless
+
 ## [1.1.1] - 2025-12-14
 
 ### Fixed